Revert "DIRECTOR: remove puppet checks when setting sprite properties"
This reverts commit bcd4a82075e92ad29ea080ade28da45dfee48896.

+++ engines/director/lingo/lingo-builtins.cpp
@@ -1658,7 +1658,7 @@ void LB::b_param(int nargs) {
 	int pos = g_lingo->pop().asInt();
 	Datum result;
 	CFrame *cf = g_lingo->_state->callstack[g_lingo->_state->callstack.size() - 1];
-	if (pos > 0 && cf->sp.argNames && cf->sp.argNames->size() <= pos) {
+	if (pos > 0 && cf->sp.argNames && (int)(cf->sp.argNames->size()) <= pos) {
 		Datum func((*cf->sp.argNames)[pos - 1]);
 		func.type = LOCALREF;
 		result = g_lingo->varFetch(func);
